+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This file contains the notable changes in the ISLa project since version 0.2a1
 
 ## [unreleased]
 
+## [1.1.0] - 2022-09-27
+
+- Added the CLI command `isla repair` and solver method `ISLaSolver.repair` for
+  repairing inputs violating semantic constraints.
+
 ## [1.0.1] - 2022-09-23
 
 ### Changed